+++ src/main/java/kitchenpos/dto/request/MenuCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.request;
 
 import java.math.BigDecimal;
 import java.util.List;
