+++ src/main/java/kitchenpos/dto/request/OrderLineItemRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.request;
 
 import kitchenpos.domain.OrderLineItem;
