+++ src/main/java/kitchenpos/dto/request/OrderTableEmptyUpdateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.request;
 
 public class OrderTableEmptyUpdateRequest {
