+++ src/main/java/kitchenpos/dto/request/TableGroupCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.request;
 
 import java.time.LocalDateTime;
 import java.util.List;
