+++ src/main/java/kitchenpos/dto/response/MenuGroupResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.response;
 
 import java.util.List;
 import java.util.stream.Collectors;
