+++ src/main/java/kitchenpos/dto/response/ProductResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.response;
 
 import java.math.BigDecimal;
 import java.util.List;
