+++ src/main/java/kitchenpos/dto/response/TableGroupResponse.java
@@ -1,4 +1,4 @@
-package kitchenpos.dto;
+package kitchenpos.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
