+++ src/test/java/kitchenpos/application/TableServiceTest.java
@@ -17,10 +17,10 @@ import kitchenpos.domain.OrderLineItem;
 import kitchenpos.domain.OrderTable;
 import kitchenpos.domain.Product;
 import kitchenpos.domain.TableGroup;
-import kitchenpos.dto.OrderTableCreateRequest;
-import kitchenpos.dto.OrderTableEmptyUpdateRequest;
-import kitchenpos.dto.OrderTableNumberOfGuestsUpdateRequest;
-import kitchenpos.dto.OrderTableResponse;
+import kitchenpos.dto.request.OrderTableCreateRequest;
+import kitchenpos.dto.request.OrderTableEmptyUpdateRequest;
+import kitchenpos.dto.request.OrderTableNumberOfGuestsUpdateRequest;
+import kitchenpos.dto.response.OrderTableResponse;
 import kitchenpos.repository.MenuGroupRepository;
 import kitchenpos.repository.MenuRepository;
 import kitchenpos.repository.OrderRepository;
